+++ src-tauri/src/commands/nostr/fetch_relays.rs
@@ -1,0 +1,17 @@
+use crate::whitenoise::Whitenoise;
+use nostr_sdk::prelude::*;
+use std::collections::HashMap;
+
+#[tauri::command]
+pub async fn fetch_relays(
+    wn: tauri::State<'_, Whitenoise>,
+) -> Result<HashMap<String, String>, String> {
+    Ok(wn
+        .nostr
+        .client
+        .relays()
+        .await
+        .into_iter()
+        .map(|(url, relay)| (url.to_string(), relay.status().to_string()))
+        .collect::<HashMap<String, String>>())
+}
