+++ src-tauri/src/commands/nostr/mod.rs
@@ -1,0 +1,33 @@
+mod decrypt_content;
+mod encrypt_content;
+mod export_nsec;
+mod fetch_contacts_with_metadata;
+mod fetch_enriched_contact;
+mod fetch_enriched_contacts;
+mod fetch_relays;
+mod generate_nip98_auth_token;
+mod init_nostr_for_current_user;
+mod invite_to_white_noise;
+mod publish_metadata;
+mod publish_relay_list;
+mod query_contacts_with_metadata;
+mod query_enriched_contact;
+mod query_enriched_contacts;
+mod search_for_enriched_contacts;
+
+pub use decrypt_content::decrypt_content;
+pub use encrypt_content::encrypt_content;
+pub use export_nsec::export_nsec;
+pub use fetch_contacts_with_metadata::fetch_contacts_with_metadata;
+pub use fetch_enriched_contact::fetch_enriched_contact;
+pub use fetch_enriched_contacts::fetch_enriched_contacts;
+pub use fetch_relays::fetch_relays;
+pub use generate_nip98_auth_token::generate_nip98_auth_token;
+pub use init_nostr_for_current_user::init_nostr_for_current_user;
+pub use invite_to_white_noise::invite_to_white_noise;
+pub use publish_metadata::publish_metadata;
+pub use publish_relay_list::publish_relay_list;
+pub use query_contacts_with_metadata::query_contacts_with_metadata;
+pub use query_enriched_contact::query_enriched_contact;
+pub use query_enriched_contacts::query_enriched_contacts;
+pub use search_for_enriched_contacts::search_for_enriched_contacts;
