+++ src-tauri/src/commands/nostr/query_contacts_with_metadata.rs
@@ -1,0 +1,20 @@
+use crate::whitenoise::Whitenoise;
+use nostr_sdk::prelude::*;
+use std::collections::HashMap;
+
+#[tauri::command]
+pub async fn query_contacts_with_metadata(
+    wn: tauri::State<'_, Whitenoise>,
+) -> Result<HashMap<String, Metadata>, String> {
+    let events = wn.nostr.query_contacts().await.map_err(|e| e.to_string())?;
+
+    let mut metadata_map = HashMap::new();
+
+    for event in events {
+        if let Ok(metadata) = serde_json::from_str::<Metadata>(&event.content) {
+            metadata_map.insert(event.pubkey.to_hex(), metadata);
+        }
+    }
+
+    Ok(metadata_map)
+}
